requirements.txt
discord.py
feedparser
beautifulsoup4